[Chore] recipient api edit
