+++ src/api/recipient.api.jsx
@@ -1,12 +1,13 @@
 import axios from "axios";
 import { showToast } from "../components/common/Toast/Toast";
 import * as C from "../constants/messageConstants";
-const BASE_URL = import.meta.env.VITE_BASE_URL;
+
+const RECIPIENTS_URL = import.meta.env.VITE_RECIPIENTS_URL;
 
 export async function deleteRecipient(id) {
   if (!id) return;
   try {
-    await axios.delete(`${BASE_URL}13-4/recipients/${id}/`);
+    await axios.delete(`${RECIPIENTS_URL}/${id}/`);
     showToast(C.TOAST_TEXT.SUCCESS_DELETE_PAGE, "success", "top");
   } catch (error) {
     showToast(C.TOAST_TEXT.FAIL_DELETE_PAGE, "error", "top");
@@ -17,7 +18,7 @@ export async function deleteRecipient(id) {
 export async function getRecipientById(id) {
   let recipient;
   try {
-    recipient = await axios.get(`${BASE_URL}13-4/recipients/${id}/`);
+    recipient = await axios.get(`${RECIPIENTS_URL}/${id}/`);
   } catch (error) {
     console.error(error);
   }
@@ -29,4 +30,25 @@ export async function getRecipientById(id) {
     recentMessages,
   } = recipient.data;
   return { name, color, img, messageCount, recentMessages };
+}
+
+export async function getAllRecipients() {
+  try {
+    const res = await axios.get(`${RECIPIENTS_URL}/?limit=100&offset=0`);
+    return res.data.results;
+  } catch (error) {
+    console.error("getAllRecipients API 에러 발생:", error.message);
+    return null;
+  }
+}
+
+export async function submitToPage(dataToSend) {
+  try {
+    const response = await axios.post(`${RECIPIENTS_URL}/`, dataToSend);
+
+    return response.data.id;
+  } catch (error) {
+    console.error("Failed to submit data:", error);
+    return null;
+  }
 }
